Update Renderable.php
You don't need to write here "magic methods". They are built-in methods in the language, you should write in an interface only the "added value" and what is MUST be in the implementation. Even if "toString" wouldn't be built-in, you don't need to mention here, because when you use the code, you never use this method from outside. Only need them.